refactor: modularize moderation service
breaks main.rs into focused modules:
- config.rs: environment config loading
- state.rs: AppState and AppError types
- auth.rs: authentication middleware
- audd.rs: AuDD fingerprinting types and handler
- xrpc.rs: ATProto labeler endpoints
- handlers.rs: health, landing, emit_label

main.rs now only composes routes (~100 lines vs ~700)

also fixes clippy warning in labels.rs (redundant closure)
